♻️refactor: Article과 ArticleList 분리 (Best에도 적용)
1파일 1컴포넌트 형식으로 변경
